feat: support for colorspaces; improvements to color extraction;
This commit adds support for color spaces - now the app should display the *exact* colors of the pixels on your screen. If you'd like to use other color spaces, those can be chosen from within the app's menu.
This also improves the calculation logic for retrieving the correct values from an NSColor (before the fence-post error was returning *very very* slightly incorrect values). Since colors on macOS are represented as Float values, there's only so much precision we can have.
We also now use a custom bitmap context to extract the color from the CGImage, this is to ensure the color we extract is in the correct color space.

Fixes #12